+++ libs/gmana/lib/forms/email.dart
@@ -1,7 +1,13 @@
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
+part of 'forms.dart';
 
 class GEmail extends StatelessWidget {
+  final TextEditingController controller;
+
+  final TextInputAction textInputAction;
+  final String hintText;
+  final List<TextInputFormatter>? inputFormatters;
+  final String? Function(String?)? validator;
+  final void Function(String)? onChanged;
   const GEmail(
     this.controller, {
     super.key,
@@ -11,13 +17,6 @@ class GEmail extends StatelessWidget {
     this.hintText = 'Enter your email',
     this.onChanged,
   });
-
-  final TextEditingController controller;
-  final TextInputAction textInputAction;
-  final String hintText;
-  final List<TextInputFormatter>? inputFormatters;
-  final String? Function(String?)? validator;
-  final void Function(String)? onChanged;
   @override
   Widget build(BuildContext context) {
     return TextFormField(