+++ libs/gmana/lib/forms/forms.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:gmana/helpers/helpers.dart';
+
+part 'email.dart';
+part 'number.dart';
+part 'password.dart';
+part 'text.dart';
+part 'validator.dart';