+++ libs/gmana/lib/g_spinners/g_spinners.dart
@@ -1,3 +1,0 @@
-export 'widgets/g_spinner.dart';
-export 'widgets/g_spinner_dot.dart';
-export 'widgets/g_spinner_wave.dart';