+++ libs/gmana/lib/g_widgets/g_widgets.dart
@@ -1,3 +1,0 @@
-export 'src/g_circular_progress.dart';
-export 'src/g_linear_progress.dart';
-export 'src/g_list_tile.dart';