+++ libs/gmana/lib/gmana.dart
@@ -1,6 +1,6 @@
 library gmana;
 
-export 'g_forms/g_forms.dart';
-export 'g_helpers/g_helpers.dart';
-export 'g_spinners/g_spinners.dart';
-export 'g_widgets/g_widgets.dart';
+export 'forms/forms.dart';
+export 'helpers/helpers.dart';
+export 'spinners/spinners.dart';
+export 'widgets/widgets.dart';