+++ libs/gmana/lib/helpers/id.dart
@@ -1,4 +1,4 @@
-import 'dart:math';
+part of 'helpers.dart';
 
 class Gid {
   final Random _random = Random();
@@ -18,8 +18,7 @@ class Gid {
   String v3(int length) {
     const ch = 'AaBbCcDdEeFfGgHhIiJjKkLlMmNnOoPpQqRrSsTtUuVvWwXxYyZz';
     final r = Random();
-    return String.fromCharCodes(
-        Iterable.generate(length, (_) => ch.codeUnitAt(r.nextInt(ch.length))));
+    return String.fromCharCodes(Iterable.generate(length, (_) => ch.codeUnitAt(r.nextInt(ch.length))));
   }
 
   String v4({int length = 8, bool hasNumbers = true, bool hasSymbols = true}) {
@@ -33,13 +32,10 @@ class Gid {
     if (hasNumbers) chars += numbers;
     if (hasSymbols) chars += symbols;
 
-    return String.fromCharCodes(Iterable.generate(
-        length, (_) => chars.codeUnitAt(r.nextInt(chars.length))));
+    return String.fromCharCodes(Iterable.generate(length, (_) => chars.codeUnitAt(r.nextInt(chars.length))));
   }
 
-  String _bits(int bitCount, int digitCount) =>
-      _printDigits(_generateBits(bitCount), digitCount);
+  String _bits(int bitCount, int digitCount) => _printDigits(_generateBits(bitCount), digitCount);
   int _generateBits(int bitCount) => _random.nextInt(1 << bitCount);
-  String _printDigits(int value, int count) =>
-      value.toRadixString(16).padLeft(count, '0');
+  String _printDigits(int value, int count) => value.toRadixString(16).padLeft(count, '0');
 }