+++ libs/gmana/lib/helpers/types.dart
@@ -1,0 +1,3 @@
+part of 'helpers.dart';
+
+typedef GMap = Map<String, dynamic>;