+++ libs/gmana/lib/spinners/delay_tween.dart
@@ -1,0 +1,13 @@
+part of 'spinners.dart';
+
+class DelayTween extends Tween<double> {
+  final double delay;
+
+  DelayTween({super.begin, super.end, required this.delay});
+
+  @override
+  double evaluate(Animation<double> animation) => lerp(animation.value);
+
+  @override
+  double lerp(double t) => super.lerp((math.sin((t - delay) * 2 * math.pi) + 1) / 2);
+}