+++ libs/gmana/lib/widgets/widgets.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart';
+
+part 'circular_progress.dart';
+part 'linear_progress.dart';
+part 'list_tile.dart';