+++ libs/gmana/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.4 (2023-12-15)
+
+- format document.
+
+## 0.0.3 (2023-10-15)
+
+- format document.
+
 ## 0.0.2 (2023-04-06)
 
 - format document.